% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/csv.R
\name{sotkanet.csv_query}
\alias{sotkanet.csv_query}
\title{Description:
retrieves Sotkanet data from the query url}
\usage{
sotkanet.csv_query(url)
}
\arguments{
\item{url}{Sotkanet CSV url}
}
\value{
sotkanet CSV query
}
\description{
Description:
retrieves Sotkanet data from the query url
}
\author{
Einari Happonen. Maintainer: Louhos \email{louhos@googlegroups.com}
}
\references{
See citation("sotkanet")
}
\keyword{utilities}

