% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json.R
\name{sotkanet.json_query}
\alias{sotkanet.json_query}
\title{Retrieves Sotkanet Data}
\usage{
sotkanet.json_query(url)
}
\arguments{
\item{url}{Sotkanet JSON url}
}
\value{
sotkanet json query
}
\description{
Retrieves Sotkanet data from the query URL.
}
\author{
Einari Happonen. Maintainer: \email{leo.lahti@iki.fi}
}
\references{
See citation("sotkanet")
}
\keyword{utilities}

