% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sotkanet_interactive.R
\name{sotkanet_interactive}
\alias{sotkanet_interactive}
\title{Interactive function for get_sotkanet}
\usage{
sotkanet_interactive(code = NULL)
}
\arguments{
\item{code}{A unique identifier for the dataset of interest.}
}
\description{
A simple interactive function, that helps with downloading Sotkanet data
}
\seealso{
[get_sotkanet()]
}
