% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sotkanet_read_cache.R
\name{sotkanet_read_cache}
\alias{sotkanet_read_cache}
\title{Read cache for sotkanet data.frame}
\usage{
sotkanet_read_cache(cache, cache_dir, query_hash)
}
\arguments{
\item{cache}{a logical whether to do caching.}

\item{cache_dir}{a path to cache directory.}

\item{query_hash}{a character used to identify the data.frame.}
}
\description{
Helper function that reads the cache for saved sotkanet data.frame.
}
\references{
See citation("sotkanet")
}
\keyword{internal}
