% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sotkanet_write_cache.R
\name{sotkanet_write_cache}
\alias{sotkanet_write_cache}
\title{Write sotkanet data.frame into cache}
\usage{
sotkanet_write_cache(cache, cache_dir, query_hash, data)
}
\arguments{
\item{cache}{a logical whether to do caching.}

\item{cache_dir}{a path to cache directory.}

\item{query_hash}{a character used to identify the data.frame.}

\item{data}{sotkanet data.frame.}
}
\description{
Helper function that writes the sotkanet data.frame into cache.
}
\references{
See citation("sotkanet")
}
\keyword{internal}
