% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SotkanetCite.R
\name{SotkanetCite}
\alias{SotkanetCite}
\title{Crate a Data Bibliography}
\usage{
SotkanetCite(id, lang = "en")
}
\arguments{
\item{id}{Indicator id.}

\item{lang}{Language for the citation. Options are English (en), Finnish (fi) and Swedish (sv).}
}
\value{
Biblatex object.
}
\description{
Crate a Data Bibliography
}
\examples{
SotkanetCite(10013, lang = "en")
SotkanetCite(10012, lang = "fi")
SotkanetCIte(10011, lang = "sv")
}
