% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SotkanetCleanCache.R
\name{SotkanetCleanCache}
\alias{SotkanetCleanCache}
\title{Clean sotkanet Cache}
\usage{
SotkanetCleanCache(cache_dir = NULL)
}
\arguments{
\item{cache_dir}{A path to the cache directory. If `NULL` (default) tries to clean default
temporary cache directory.}
}
\description{
Delete all .rds files from the sotkanet cache directory.
}
\examples{
\dontrun{
SotkanetCleanCache()
}
}
