% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{SotkanetCollect}
\alias{SotkanetCollect}
\title{Convert Sotkanet Data from list to data.frame}
\usage{
SotkanetCollect(x, name)
}
\arguments{
\item{x}{input data (from SotkanetIndicators or SotkanetRegions etc.)}

\item{name}{name for the column ("indicator", "region", etc.)}
}
\value{
sotkanet data table
}
\description{
Converts Sotkanet list object to data.frame.
}
\author{
Einari Happonen. Maintainer: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("sotkanet")
}
\keyword{utilities}

