% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SotkanetData.R
\name{SotkanetData}
\alias{SotkanetData}
\title{Retrieve Sotkanet Daata.}
\usage{
SotkanetData(indicator, years, genders)
}
\arguments{
\item{indicator}{Dataset identifier}

\item{years}{vector of years c(2010, 2012, ... )}

\item{genders}{vector of genders ('male' | 'female' | 'total')}
}
\value{
sotkanet json query
}
\description{
Retrieve Sotkanet data according to the query arguments.
}
\examples{
\dontrun{d <- SotkanetData(indicator, years, genders)}
}
\author{
Einari Happonen. Maintainer: Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("sotkanet")
}
\keyword{utilities}

