% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SotkanetIndicatorMetadata.R
\name{SotkanetIndicatorMetadata}
\alias{SotkanetIndicatorMetadata}
\title{Sotkanet Indicator Metadata}
\usage{
SotkanetIndicatorMetadata(id = NULL, flatten = TRUE)
}
\arguments{
\item{id}{Indicator id}

\item{flatten}{Parameter passed to function handling json files, default 
is TRUE}
}
\value{
sotkanet indicator metadata as a list object
}
\description{
Retrieves sotkanet indicator metadata.
}
\details{
Data is fetched from \preformatted{https://sotkanet.fi/rest/1.1/indicators/<id>}
}
\examples{
\dontrun{
x <- SotkanetIndicatorMetadata(10013)
}
}
\references{
See citation("sotkanet")
}
\author{
Maintainer: Pyry Kantanen
}
\keyword{utilities}
