% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SotkanetIndicators.R
\name{SotkanetIndicators}
\alias{SotkanetIndicators}
\title{Sotkanet Indicators}
\usage{
SotkanetIndicators(id = NULL, type = "table")
}
\arguments{
\item{id}{Dataset identifier. Default is NULL returning all}

\item{type}{Parameter passed to function handling json files, default is TRUE}
}
\value{
json query in data.frame
}
\description{
Retrieve Sotkanet indicator metadata
}
\details{
Data is fetched from \url{https://sotkanet.fi/rest/1.1/indicators}.
}
\examples{
\dontrun{
sotkanet.indicators <- SotkanetIndicators(type = "table")
}
}
\references{
See citation("sotkanet")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}, Pyry Kantanen
}
\keyword{utilities}
