% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SotkanetInteractive.R
\name{SotkanetInteractive}
\alias{SotkanetInteractive}
\title{Interactive function for GetDataSotkanet}
\usage{
SotkanetInteractive(code = NULL)
}
\arguments{
\item{code}{A unique identifier for the dataset of interest.}
}
\description{
A simple interactive function, that helps with downloading Sotkanet data
}
\seealso{
[GetDataSotkanet()]
}
