% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_funs.R
\name{try_GET}
\alias{try_GET}
\title{HTTP function: try_GET}
\source{
Gracefully failing HTTP request code (slightly adapted by Pyry 
   Kantanen) from RStudio community member kvasilopoulos. Many thanks!
   
   Source of the original RStudio community discussion:
   \url{https://community.rstudio.com/t/internet-resources-should-fail-gracefully/49199}
}
\usage{
try_GET(url = NULL, timeout = 10)
}
\arguments{
\item{url}{URL to query}

\item{timeout}{Timeout time in seconds.}
}
\value{
HTTP response or a conditionMessage
}
\description{
Constructs a function for trying an URL
}
\author{
Pyry Kantanen (maintainer), Kostas Vasilopoulos
}
\keyword{internal}
