% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_frictionless_metadata.R
\name{write_frictionless_metadata}
\alias{write_frictionless_metadata}
\title{Writes the sotkanet data.frame in datapackage form}
\usage{
write_frictionless_metadata(indicators, data)
}
\arguments{
\item{indicators}{Dataset identifier}

\item{data}{The sotkanet data.frame object}
}
\value{
datapackage
}
\description{
A function that turns the data.frame into a datapackage with the metadata added in.
}
\references{
See citation("sotkanet")
}
\keyword{internal}
