% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/base_url.R
\name{base_url}
\alias{base_url}
\title{Base url to the Sotkanet API}
\usage{
base_url(...)
}
\arguments{
\item{...}{Arguments to be passed

Returns:}
}
\value{
sotkanet API base url
}
\description{
Arguments:
}
\examples{
base_url()
}
\author{
Leo Lahti \email{louhos@googlegroups.com}
}
\references{
See citation("sotkanet")
}
\keyword{utilities}

