% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_url.R
\name{base_url}
\alias{base_url}
\title{Sotkanet Base URL}
\usage{
base_url(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
sotkanet API base url
}
\description{
Base url for sotkanet api.
}
\examples{
base_url()
}
\references{
See citation("sotkanet")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
