% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetDataSotkanet.R
\name{GetDataSotkanetSingleIndicator}
\alias{GetDataSotkanetSingleIndicator}
\title{Description:
GetDataSotkanetSingleIndicator retrieves Sotkanet data
for given indicator according to the query arguments and combines
indicator, region, and overall data into one table}
\usage{
GetDataSotkanetSingleIndicator(indicator, years = 1990:2000,
  genders = "total", regions = NULL, region.category = NULL)
}
\arguments{
\item{indicator}{Dataset identifier}

\item{years}{vector of years c(2010, 2012, ... )}

\item{genders}{vector of genders ('male' | 'female' | 'total')}

\item{regions}{return selected regions only}

\item{region.category}{return selected regions category (for options, see:
         unique(SotkanetRegions(type = "table")$region.category));
        "ALUEHALLINTOVIRASTO, "ERVA", "EURALUEET", "EUROOPPA", "KUNTA",
        "MAA", "MAAKUNTA", "NUTS1", "POHJOISMAAT", "SAIRAANHOITOPIIRI",
        "SEUTUKUNTA", "SUURALUE"

Returns:}
}
\value{
sotkanet data table
}
\description{
Arguments:
}
\examples{
#
}
\author{
Einari Happonen. Maintainer: Louhos/Opasnet \email{louhos@googlegroups.com}
}
\references{
See citation("sotkanet")
}
\keyword{utilities}

