\docType{package}
\name{sotkanet-package}
\alias{sotkanet}
\alias{sotkanet-package}
\title{R Tools for Sotkanet API}
\description{
Brief summary of the sotkanet package
}
\details{
\tabular{ll}{ Package: \tab sotkanet\cr Type: \tab
Package\cr Version: \tab See sessionInfo() or DESCRIPTION
file\cr Date: \tab 2013\cr License: \tab FreeBSD\cr
LazyLoad: \tab yes\cr }

R Tools for Sotkanet API
}
\examples{
library(sotkanet)
}
\author{
Leo Lahti, Einari Happonen, Juuso Parkkinen and Joona
Lehtomaki \email{louhos@googlegroups.com}
}
\references{
See citation("sotkanet")
\url{http://ropengov.github.com/sotkanet}
}
\keyword{package}

