% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sotkanet-package.R
\docType{package}
\name{sotkanet-package}
\alias{sotkanet-package}
\alias{sotkanet}
\title{Tools for Sotkanet API}
\description{
Retrieval of sotkanet data.

\tabular{ll}{
  Package: \tab sotkanet\cr
  Type: \tab Package\cr
  Version: \tab See sessionInfo() or DESCRIPTION file\cr
  Date: \tab 2013-2017\cr
  License: \tab FreeBSD\cr
  LazyLoad: \tab yes\cr
}
}
\examples{
library(sotkanet)
}
\references{
See citation("sotkanet")
}
\author{
Leo Lahti, Einari Happonen, Juuso Parkkinen and Joona Lehtomaki \email{leo.lahti@iki.fi}
}
\keyword{package}
