\name{sotkanet.csv_query}
\alias{sotkanet.csv_query}
\title{Description:
retrieves Sotkanet data from the query url}
\usage{
sotkanet.csv_query(url)
}
\arguments{
  \item{url}{Sotkanet CSV url

  Returns:}
}
\value{
sotkanet CSV query
}
\description{
Arguments:
}
\author{
Einari Happonen / Opasnet. Maintainer: Louhos/Opasnet
\email{louhos@googlegroups.com}
}
\references{
See citation("sotkanet")
}
\keyword{utilities}

