% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{sotkanet.csv_query}
\alias{sotkanet.csv_query}
\title{Retrieve Sotkanet Data (csv)}
\usage{
sotkanet.csv_query(url, user.agent = NULL, ...)
}
\arguments{
\item{url}{Sotkanet CSV url}

\item{user.agent}{User agent defined by the user. Default is 
"rOpenGov/sotkanet"}

\item{...}{additional parameters to be passed to test_connection, mainly 
timeout (in seconds, default is 10)}
}
\value{
data.frame
}
\description{
Retrieve data in csv format from given url.
}
\references{
See citation("sotkanet")
}
\author{
Maintainer: Leo Lahti \email{leo.lahti@iki.fi}, Pyry Kantanen
}
\keyword{internal}
