% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SotkanetCite.R
\name{SotkanetCite}
\alias{SotkanetCite}
\title{Crate a Data Bibliography}
\usage{
SotkanetCite(id)
}
\arguments{
\item{id}{Indicator id}
}
\value{
Biblatex object
}
\description{
Crate a Data Bibliography
}
\examples{
SotkanetCite(10013)
}
