\name{SotkanetCollect}
\alias{SotkanetCollect}
\title{Description:
SotkanetCollect converts the list object from Sotkanet to a data.frame}
\usage{
SotkanetCollect(x, name)
}
\arguments{
  \item{x}{input data (from SotkanetIndicators or
  SotkanetRegions etc.)}

  \item{name}{name for the column ("indicator", "region",
  etc.)

  Returns:}
}
\value{
sotkanet data table
}
\description{
Arguments:
}
\author{
Einari Happonen / Opasnet. Maintainer: Louhos
\email{louhos@googlegroups.com}
}
\references{
See citation("sotkanet")
}
\keyword{utilities}

