% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/internal.R
\name{SotkanetData}
\alias{SotkanetData}
\title{Description:
SotkanetData retrieves Sotkanet data
according to the query arguments.}
\usage{
SotkanetData(indicator, years, genders)
}
\arguments{
\item{indicator}{Dataset identifier}

\item{years}{vector of years c(2010, 2012, ... )}

\item{genders}{vector of genders ('male' | 'female' | 'total')

Returns:}
}
\value{
sotkanet json query
}
\description{
Arguments:
}
\examples{
#
}
\author{
Einari Happonen / Opasnet / Louhos. Maintainer: Louhos/Opasnet \email{louhos@googlegroups.com}
}
\references{
See citation("sotkanet")
}
\keyword{utilities}

