% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SotkanetIndicatorMetadata.R
\name{SotkanetIndicatorMetadata}
\alias{SotkanetIndicatorMetadata}
\title{Sotkanet Indicator Metadata. (old version)}
\usage{
SotkanetIndicatorMetadata(id = NULL, type = "raw", user.agent = NULL)
}
\arguments{
\item{id}{Indicator id}

\item{type}{Parameter passed onto \code{\link{SotkanetIndicators}}. Default
is 'raw' for the whole output but 'table' is also supported}

\item{user.agent}{"User agent" defined by the user. Default is NULL which
will use the default package identifier "rOpenGov/sotkanet"}
}
\value{
sotkanet indicator metadata as a list object
}
\description{
Retrieves sotkanet indicator metadata. (This is an older version of the function. 
 It is advised to use the new [sotkanet_indicator_metadata()] function instead.)
}
\details{
Data is fetched from \preformatted{https://sotkanet.fi/rest/1.1/indicators/<id>}
}
\examples{
\dontrun{
x <- SotkanetIndicatorMetadata(10013)
}
}
\references{
See citation("sotkanet")
}
\author{
Maintainer: Pyry Kantanen
}
\keyword{utilities}
