% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SotkanetIndicators.R
\name{SotkanetIndicators}
\alias{SotkanetIndicators}
\title{Sotkanet Indicators}
\usage{
SotkanetIndicators(id = NULL, type = "table")
}
\arguments{
\item{id}{Dataset identifier. Default is NULL returning all}

\item{type}{type output format, either 'table' (default) or 'raw'. 
Default produces a truncated table with strictly defined columns that
are useful in other functions. 'Raw' produces the full output which might
be useful for exploratory purposes.}
}
\value{
data.frame or a list
}
\description{
Retrieve Sotkanet indicator metadata
}
\details{
Data is fetched from \url{https://sotkanet.fi/rest/1.1/indicators}.
}
\examples{
\dontrun{
sotkanet.indicators <- SotkanetIndicators(type = "table")
}
}
\references{
See citation("sotkanet")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}, Pyry Kantanen
}
\keyword{utilities}
