% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SotkanetIndicators.R
\name{SotkanetIndicators}
\alias{SotkanetIndicators}
\title{Sotkanet Indicators}
\usage{
SotkanetIndicators(id = NULL, type = "table")
}
\arguments{
\item{id}{Dataset identifier}

\item{type}{output format ("table" or "raw")}
}
\value{
json query in selected output format
}
\description{
Retrieve data for a specific data id.
  \url{https://sotkanet.fi/rest/1.1/indicators}.
}
\examples{
\dontrun{sotkanet.indicators <- SotkanetIndicators(type = "table")}
}
\references{
See citation("sotkanet")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
