% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SotkanetIndicators.R
\name{SotkanetIndicators}
\alias{SotkanetIndicators}
\title{SotkanetIndicators}
\usage{
SotkanetIndicators(id = NULL, type = "table")
}
\arguments{
\item{id}{Dataset identifier}

\item{type}{output format ("table" or "raw")}
}
\value{
sotkanet json query in selected output format
}
\description{
SotkanetIndicators retrieves Sotkanet data corresponding to a
specified data identifier from
\url{http://www.sotkanet.fi/rest/1.1/indicators}
}
\examples{
\dontrun{sotkanet.indicators <- SotkanetIndicators(type = "table")}
}
\author{
Einari Happonen. Maintainer: Louhos/Opasnet \email{louhos@googlegroups.com}
}
\references{
See citation("sotkanet")
}
\keyword{utilities}

