% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sotkanet.R
\name{SotkanetRegions}
\alias{SotkanetRegions}
\title{Description:
SotkanetRegions retrieves Sotkanet regions data from
\url{http://www.sotkanet.fi/rest/1.1/regions}}
\usage{
SotkanetRegions(type = "table")
}
\arguments{
\item{type}{Return format ("table" or "raw")

Returns:}
}
\value{
sotkanet json query in selected format
}
\description{
Arguments:
}
\examples{
# sotkanet.regions <- SotkanetRegions(type = "table")
}
\author{
Einari Happonen / Opasnet / Louhos. Maintainer: Louhos/Opasnet \email{louhos@googlegroups.com}
}
\references{
See citation("sotkanet")
}
\keyword{utilities}

