% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SotkanetRegions.R
\name{SotkanetRegions}
\alias{SotkanetRegions}
\title{Sotkanet Regions}
\usage{
SotkanetRegions(type = "table", user.agent = NULL)
}
\arguments{
\item{type}{type output format, either 'table' (default) or 'raw'. 
Default produces a truncated table with strictly defined columns that
are useful in other functions. 'Raw' produces the full output which might
be useful for exploratory purposes.}

\item{user.agent}{"User agent" defined by the user. Default is NULL which
will then use the package identifier "rOpenGov/sotkanet"}
}
\value{
data.frame
}
\description{
Retrieves sotkanet regions data.
}
\details{
Data is fetched from \url{https://sotkanet.fi/rest/1.1/regions}.
}
\examples{
\dontrun{
sotkanet.regions <- SotkanetRegions(type = "table")
}
}
\references{
See citation("sotkanet")
}
\author{
Maintainer: Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
