% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sotkanet.R
\name{SotkanetRegions}
\alias{SotkanetRegions}
\title{Sotkanet Regions}
\usage{
SotkanetRegions(type = "table")
}
\arguments{
\item{type}{Return format ("table" or "raw")}
}
\value{
sotkanet json query in selected format
}
\description{
Retrieves sotkanet regions data.
}
\details{
Data is fetched from \url{http://www.sotkanet.fi/rest/1.1/regions}.
}
\examples{
\dontrun{sotkanet.regions <- SotkanetRegions(type = "table")}
}
\references{
See citation("sotkanet")
}
\author{
Maintainer: Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
